+++ CHANGELOG.md
@@ -11,10 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Make `as_limbs_mut` const ([#757])
 - Impl `RemMixed` for even splits as well ([#791])
-
+- Make `Integer` be fmt::Binary, LowerHex and UpperHex ([#792])
+- 
 [#757]: https://github.com/RustCrypto/crypto-bigint/pull/757
 [#760]: https://github.com/RustCrypto/crypto-bigint/pull/760
 [#791]: https://github.com/RustCrypto/crypto-bigint/pull/791
+[#792]:https://github.com/RustCrypto/crypto-bigint/pull/792
 
 ## 0.6.0 (2025-01-22)
 ### Added